pyproject.toml

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prospect-automation"
version = "0.1.0"
description = "Prospect Research Automation 2.1 (EB-focused)"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name="Ethan", email="example@example.com"}]
dependencies = [
    "pandas>=2.2.2",
    "typer>=0.12.3",
    "rich>=13.7.1",
    "pydantic>=2.8.2",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "rapidfuzz>=3.9.4",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.2",
    "tldextract>=5.1.2",
    "matplotlib>=3.9.0",
    "openpyxl>=3.1.5",
    "fpdf2>=2.8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
    "black>=24.8.0",
    "isort>=5.13.2",
    "ruff>=0.6.9",
]

[project.scripts]
prospect = "prospect_automation.cli:app"
